Add function to server script to extract HTML accept header
Two of the idiosyncratic pattern matching functions for ABCD depended on exact matching of the Accept header with "text/html". But browsers that send Accept headers with multiple possible media types were not matching. A function was added that checks all accept headers for a match to "text/html" was added and applied to the Accept headers sent by browsers.