changes to source files
It isn't necessary to include CSVs for namespaces that don't have changes, so I deleted them as well as the references to them in the config.json file. I think if they are included, it will spawn new versions of their term lists that don't actually include any changes. So it's probably better to just leave them out.

Also, in the config.json file, the modifications_file_path has to match the actual name of the corresponding CSV file. So I changed that to terms_2023-07-07 to match.